+++ Plugins/Effect/Effect.hpp
@@ -2,6 +2,7 @@
 
 #include "Plugin.hpp"
 #include "Services/Events/Events.hpp"
+#include "Services/Hooks/Hooks.hpp"
 #include "API/Types.hpp"
 
 using ArgumentStack = NWNXLib::Services::Events::ArgumentStack;
@@ -17,7 +18,7 @@ public:
 private:
     ArgumentStack PackEffect(ArgumentStack&& args);
     ArgumentStack UnpackEffect(ArgumentStack&& args);
-
+    ArgumentStack SetOnEffectRemovedScript(ArgumentStack&& args);
 };
 
 }